[MOB-16746] Refactor Sample App to Test App and refactored schemes and targets
