+++ AEPMedia.xcodeproj/xcshareddata/xcschemes/AEPMediaFunctionalTests.xcscheme
@@ -15,8 +15,8 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "D66F0B642616929C0018E779"
-               BuildableName = "AEPMediaFunctionalTests.xctest"
-               BlueprintName = "AEPMediaFunctionalTests"
+               BuildableName = "FunctionalTests.xctest"
+               BlueprintName = "FunctionalTests"
                ReferencedContainer = "container:AEPMedia.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
@@ -44,8 +44,8 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "D66F0B642616929C0018E779"
-               BuildableName = "AEPMediaFunctionalTests.xctest"
-               BlueprintName = "AEPMediaFunctionalTests"
+               BuildableName = "FunctionalTests.xctest"
+               BlueprintName = "FunctionalTests"
                ReferencedContainer = "container:AEPMedia.xcodeproj">
             </BuildableReference>
          </TestableReference>
@@ -72,8 +72,8 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "D66F0B642616929C0018E779"
-            BuildableName = "AEPMediaFunctionalTests.xctest"
-            BlueprintName = "AEPMediaFunctionalTests"
+            BuildableName = "FunctionalTests.xctest"
+            BlueprintName = "FunctionalTests"
             ReferencedContainer = "container:AEPMedia.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
