+++ AEPMedia.xcodeproj/xcshareddata/xcschemes/AEPMediaUnitTests.xcscheme
@@ -15,8 +15,8 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2EFF1B5925BF57FC00375965"
-               BuildableName = "AEPMediaUnitTests.xctest"
-               BlueprintName = "AEPMediaUnitTests"
+               BuildableName = "UnitTests.xctest"
+               BlueprintName = "UnitTests"
                ReferencedContainer = "container:AEPMedia.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
@@ -44,8 +44,8 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2EFF1B5925BF57FC00375965"
-               BuildableName = "AEPMediaUnitTests.xctest"
-               BlueprintName = "AEPMediaUnitTests"
+               BuildableName = "UnitTests.xctest"
+               BlueprintName = "UnitTests"
                ReferencedContainer = "container:AEPMedia.xcodeproj">
             </BuildableReference>
          </TestableReference>
@@ -72,8 +72,8 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2EFF1B5925BF57FC00375965"
-            BuildableName = "AEPMediaUnitTests.xctest"
-            BlueprintName = "AEPMediaUnitTests"
+            BuildableName = "UnitTests.xctest"
+            BlueprintName = "UnitTests"
             ReferencedContainer = "container:AEPMedia.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
