+++ Samples/TestApp/README.md
@@ -1,0 +1,11 @@
+# How to run Sample App
+
+Go to Terminal:
+1. cd aepsdk-media-ios/build
+2. pod install 
+3. pod update
+4. Open AEPMedia.xcworkspace in xcode
+5. Choose MediaSampleApp Scheme 
+6. In MediaSampleApp > AppDelegate.swift, provide launch app id in 
+   private let LAUNCH_ENVIRONMENT_FILE_ID = "your launch app id"
+7. Run the app
